# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Headmaster API - Personal AI Knowledge System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inputs/youtube:
    post:
      operationId: fetchYoutubePlaylist
      tags: [inputs]
      summary: Fetch a YouTube video or playlist and create an input from its content
      description: |
        For single video URLs this endpoint returns a regular JSON response (HTTP 201,
        Content-Type: application/json) matching the YoutubeImportResult schema.

        For playlist URLs (containing `list=` query parameter) the endpoint returns an
        HTTP 200 NDJSON streaming response (Content-Type: application/x-ndjson).
        Each newline-delimited JSON object is one of:
          - `{"type":"start","total":N,"playlistTitle":"..."}` — emitted once before processing begins
          - `{"type":"video","index":N,"total":N,"title":"...","input":{...},"hasTranscript":true}` — one created input
          - `{"type":"skip","index":N,"total":N,"title":"...","reason":"..."}` — unavailable/private video skipped
          - `{"type":"error","message":"..."}` — internal failure; stream ends immediately after
          - `{"type":"done","added":N,"skipped":N}` — emitted once all videos are processed

        Clients importing playlists should consume this stream directly rather than
        using the generated JSON client hook.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/YoutubePlaylistBody"
      responses:
        "200":
          description: Playlist import stream (application/x-ndjson) — see description for event schema
        "201":
          description: Input created from single video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YoutubeImportResult"
...
components:
  schemas:
...
    Input:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [text, url, note, file, youtube, instagram, pinterest, image, video, audio]
        summary:
          type: string
          nullable: true
        processed:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - content
        - type
        - processed
        - createdAt
        - updatedAt
...
    PlaylistEntry:
      type: object
      properties:
        title:
          type: string
        videoId:
          type: string
          nullable: true
        hasTranscript:
          type: boolean
      required:
        - title
        - hasTranscript
    YoutubeImportResult:
      type: object
      properties:
        input:
          $ref: "#/components/schemas/Input"
        playlistEntries:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/PlaylistEntry"
      required:
        - input
    YoutubePlaylistBody:
      type: object
      properties:
        url:
          type: string
          description: Full YouTube video or playlist URL (e.g. youtube.com/watch?v=xxx or youtube.com/playlist?list=PLxxx)
        title:
          type: string
          description: Optional custom title (defaults to the title from YouTube)
        fetchTranscripts:
          type: boolean
          description: When true (default), fetches a transcript for each video in a playlist. Set to false to skip transcript fetching and speed up import.
      required:
        - url
...